// streamlit_molstar/pocket/frontend/src/custom-types.ts
export interface RegionData {
    name: string;
    start: number;
    end: number;
}

export interface ScoresData {
    conservation?: number[];
    plddt?: number[]; //also referenced to as AlphaFold score
}

export interface StructureData {
    indices: string[];
    sequence: string[];
    binding: number[];
    regions: RegionData[];
    scores: ScoresData;
}

export interface PocketData {
    name: string;
    rank: string;
    score: string;
    probability: string;
    center: string[];
    residues: string[];
    surface: string[];
    color?: string;             //color of the pocket, if already generated by rcsb plugin (e.g. "00ff00")
    isVisible?: boolean;        //if the pocket is visible
    avgConservation?: number;   //computed average conservation of the pocket
    avgAlphaFold?: number;      //computed average AlphaFold score of the pocket
}

export interface Metadata {
    //metadata can contain anything at this point
}

export interface PredictionData {
    structure: StructureData;
    pockets: PocketData[];
    metadata: Metadata;
}
// ...
export enum PocketsViewType {
    Ball_Stick_Atoms_Color = 0,
    Ball_Stick_Residues_Color = 1,
    Surface_Atoms_Color = 2,
    Surface_Residues_Color = 3
}